fix(ci): Python virtualenv and WASM bigint type
- Create virtualenv before maturin develop in Python CI
- Change setClockSkewToleranceSeconds to accept i32 (JS number) instead
  of i64 (JS bigint) for better JavaScript ergonomics

+++ core/claim169-wasm/src/lib.rs
@@ -284,8 +284,8 @@ impl WasmDecodeOptions {
 
     /// Set clock skew tolerance in seconds for timestamp validation
     #[wasm_bindgen(js_name = "setClockSkewToleranceSeconds")]
-    pub fn set_clock_skew_tolerance_seconds(mut self, seconds: i64) -> WasmDecodeOptions {
-        self.clock_skew_tolerance_seconds = seconds.max(0);
+    pub fn set_clock_skew_tolerance_seconds(mut self, seconds: i32) -> WasmDecodeOptions {
+        self.clock_skew_tolerance_seconds = i64::from(seconds).max(0);
         self
     }
 }
